#1341 - Merge Configurable trait into Bootstrap class
